requirements.txt
Flask
Flask-WTF
Flask-Login
Flask-SQLAlchemy
Flask-Session
Flask-Limiter
python-dotenv
Flask-Talisman
WTForms
